Generate smart constructors signature and "verify" implementation
Summary:
This is first step for the smart constructor implementation.

We need to spot places where make_* functions are called not in the correct order. If we generate syntax-tree based on the order of the make_* function calls, and then compare it with the correct one, we'll be able to locate places with incorrect invactions. This is exactly what we are doing here.

This change introduces generator for smart constructor set signature and one implementation that is based on the "stack"-like state passed everywhere.

Next step will involve changing all the invocations of the make_* functions to pass `parser` and to return one. This is where we'll put our smart constructors and a state.

Differential Revision: D6770015

fbshipit-source-id: ae4b9c72f46bbf799bbba997973306878427c834