Record hackc-compiled files in the static contents cache
Summary:
Various parts of url-resolving, and the final decision to execute a
url depend on being able to test for the existence of a php file in
the repo.

Historically we've used the static contents file for this; at some
point we should investigate inspecting the repo directly instead - but
for now just make sure that all the parsed files end up in static
contents.

Reviewed By: alexeyt

Differential Revision: D6814622

fbshipit-source-id: 337afcae30360af6b05b9e49b2c8642554f7b96f