allow ? as first character of ID in lexer
Summary:
hhas files can apparently have question marks at the start of
arguments, as in AssertRATL $argc ?Int. We extend the lexer to allow ?
as the first character of an ID token. This is far too generous, as ID
is used in various other places where ? certainly shouldn't be
allowed, but I think it's OK for parsing well-formed inputs.

Reviewed By: vladima, gregory-nisbet

Differential Revision: D6821726

fbshipit-source-id: 13a3d523c2e5b80ba19bec4abb9287dee8e0529b