Remove AutocompleteService's use of the naming lvar hook
Summary:
After {D6756078}, local variable names are available during typechecking, so it is not necessary for AutocompleteService to attach a naming hook to track them.

Instead of suggesting local identifiers tracked in the naming hook, in this change we add a function to retrieve Typing_env's map from local identifiers to types, and use it in the autocomplete typing hook. The typechecker added `$this` as a local in static methods (seemingly unnecessarily), so we check `is_static` before adding a `$this` local. Since `$this` is available in the mapping in instance methods, we also remove the special-casing for `$this` in AutocompleteService.

Reviewed By: pittsw

Differential Revision: D6811056

fbshipit-source-id: 4c0d103f14cd809b8cc11748f4a417eb5f91194a