Kill Object::HasNativeData; add HeaderKind::NativeObject
Summary:
All HNI classes as well as generator and async generator set
HasNativeData, and have a NativeData prefix that holds the Cpp data.
Introduce HeaderKind::NativeObject, and kill the attribute.

This actually simplifies a few things; when we're switching on
HeaderKind in a context that cares about native data, the extra
case lets us do the right thing for NativeObject (has data) and
Object (doesn't have data) without an extra check.

HasInstanceDtor is now simpler; it's range check now includes
the new HeaderKind.

Reviewed By: jano, alexeyt

Differential Revision: D6800425

fbshipit-source-id: 85327bfe3fb8a1c564132c0caeca695c64b51b38