Move ObjectData::CallToImpl to Class::Attributes
Summary:
It doesn't vary by instance.

In order for this to be set at class-init time for SimpleXMLElement,
which can be extended by other classes, bring back NativeDataInfo::cls_attrs,
and check it when initializing the class. This ensures subclasses pick
it up as well, and avoids an extra load/or/store when initializing
every SimpleXMLElement/Iterator instance.

This flag is also set for collection classes, which has a different
(non-HNI) initialization path.

Reviewed By: alexeyt

Differential Revision: D6800572

fbshipit-source-id: 667f9be459ceed83862ae3f48975eca22bc1866b