Add <<__PHPStdLib>> to builtins
Summary:
Add `<<__PHPStdLib>>` to remaining non-magic functions in `.hhi` files. I will then checkin a proxy of these files in WWW. This proxy hhi file that's checked into WWW

WHY?

Because this will allow me to blacklist these builtins as they are removed from WWW, atomically with a single WWW diff. This is important because having to wait a week for Hack to push and then make sure it's clean every time is time consuming.

Reviewed By: jamesjwu

Differential Revision: D6821752

fbshipit-source-id: 7bce88705aaec71fe502f622a9497bc1a4210789