Fix environment threading in default value printing
Summary:
Fix environment threading in default value printing

Prior to this diff, `use_single_quote` flag was only making it to the top level and was not being passed to the subcalls. Adding this to each call site would be excessive amount of code, hence this diff creates a default value printing environment so minimize extra code.

Reviewed By: vladima

Differential Revision: D6847506

fbshipit-source-id: a00012732546bfa7d6cfea88f5cc48045809b1de