Optimize GC trigger check
Summary:
Trim the meat of the check down to one comparison by setting m_nextGC
to INT64_MAX when the GC is disabled and moving the rds::header() check to
resetGC(), which happens just a handful of times per request. A call to
checkGC() now only adds two instructions to the hot path.

Reviewed By: edwinsmith

Differential Revision: D6797196

fbshipit-source-id: bdc3a274d1485505a73439c39e94e556f59f9dad