Type check lambdas at definition site if used in untyped context
Summary: If a lambda is passed as an untyped parameter, or returned as an untyped result, let's check its definition with no types for its parameters, rather than defaulting to the legacy Tanon technique. There is little observable change (no new tests) but we get to eliminate another class of lambdas from the legacy technique.

Reviewed By: dlreeves

Differential Revision: D6844952

fbshipit-source-id: c287bc92f3c07dc7e73fc080e8ec88a1d044c7fe