Add strongly-typed LSP messages
Summary:
Previously, we'd been carrying around JSON for incoming LSP
messages, and only parsing them at the moment of need. This diff
introduces strongly-typed ADTs for all possible incoming LSP messages,
and an up-front parser for them.

(It's a bit complicated because there's no way to parse an LSP response
without knowing what kind of response is expected, i.e. by knowing what
request it is in response to. So I've added that).

I've also included ADTs for outgoing LSP messages as well.

Reviewed By: arxanas

Differential Revision: D6851537

fbshipit-source-id: d501df466b5af53a3cae47bd52b0396d8d2fda2e