Make utils/file_content standalone, v2
Summary:
utils/file_content is a library for relating line+column-based positions to offsets in a buffer, and for applying edits to that buffer.

Currently it's tied to ide_api_types, part of Hack's API for communicating over its persistent connection. But it's more fundamental than that. It's just a general-purpose need to have a buffer and track incremental updates. In particular, both `flow lsp` and the flow server itself need to do this.

This diff creates new types 'position, range, text_edit' for use inside file_content.ml. The idea is that these primitive types should be used internally, including by Flow.

It also retains the pre-existing types 'position, range, text_edit' inside Ide_api_types.ml -- the idea is that these types are used for the wire format.

(I experimented with a different diff, D6489909, which got rid of the Ide_api_types versions entirely, and switched everything over to the File_content types. That involved considerable churn. And in the end I think it's legit and reasonable to have separate types for the "internal operations" vs the "wire format". LSP already does that too).

Reviewed By: gabelevi

Differential Revision: D6506951

fbshipit-source-id: 5ac95ca90b967d5c352a5407da1ac73f4338607a