Legacy bug fix; re-align lexer before skipping statement
Summary:
When an unexpected `Tword` token occurs at the top-level when trying to skip
over ignored parts of a parse, the legacy parser didn't put that unexpected
`Tword` back in the input. This leads to `ignore_statement` sometimes not
correctly ignoring the remainder. This only happens in PHP files, so the type
checker wasn't affected, but for parity comparisson with the new front-end,
this gets very in the way.

Reviewed By: andrewjkennedy

Differential Revision: D6870681

fbshipit-source-id: d2920cc6ac9dbefa2abf3b54f54cc01d66c6512e