Legacy fix; do not promote ignored methods to functions
Summary:
When the legacy parser parses files in PHP mode, and it comes across a
`var`-denoted member variable declaration in a class, it starts skipping the
remainder of the class. By not recognising the keywords `public` and `private`,
the parser incorrectly concludes that the next thing is a function declaration
//outside of the class definition//. For parity comparisson with the new
front-end this causes a //lot// of noise.

Reviewed By: andrewjkennedy

Differential Revision: D6870680

fbshipit-source-id: e4e3a4c7f67bdf505bdf3c211871be97be78bf73