Add support for repo-auth-types for Hack arrays
Summary:
Add full support for repo-auth-types for Hack arrays. This is mainly mechanical
changes to the emission and parsing code, as well as removing a few
assertions. Dict and Keyset can technically get repo-auth-types (even though
they only represent packed data right now) if they happen to be
"vector-like". This is very rare however. Utilize this information in a few
places in the JIT for optimization.

The (disabled) logic in tvMatchesArrayType() was incorrect when it tried to
verify if an array was packed. Fix that and simplify the function a bit. This
turned up that SetArray::IsVectorData() was incorrect (it always returned
false), so implement that properly.

Finally, enable the logic in HHBBC to emit PackedN repo-auth-types. We can
actually use them in the JIT (for inferring element types, for example), and it
doesn't seem to hurt anything to emit them.

Reviewed By: markw65

Differential Revision: D6715443

fbshipit-source-id: ccb276291032ae18dc1f5402e3044661c8c683d6