Make the failure for missing whitespace after async less absolute
Summary:
For HHVM compatibility reasons, we failed hard on `async()` (no whitespace).
Other users of the full fidelity front-end shouldn't. Also, this was ported
before we had a `hhvm_compat_mode` and a post-parsing error checker. This diff
moves the hard failure to the right place and only triggers it when in
`hhvm_compat_mode`.

Reviewed By: arxanas

Differential Revision: D6873329

fbshipit-source-id: b14154437dec961bbee03643a21047a7013d7dbd