MultiWorker.call_with_interrupt
Summary:
The main goal of this stack is to be able to cancel a lengthy typechecking / redeclaration job, and do it fast enough to make it usable in IDE scenarios (i.e make "cancel job + process an edit + answer autocomplete request" combo responsive).

The approach I implement here is pretty simple - I set a flag inside shared memory that is checked in a bunch of core functions (`hh_add`, `hh_get`, etc...), and make workers exit when they see it. The other things that I tried were:

- just waiting for current batch to stop (avoiding need to deal with half-completed jobs)
- signal based approach. They require first [extracting slave pid to master](differential/diff/41711009/) (so the master knows who to signal), and then ensuring that the signal is not delivered in the middle of atomic operations in `hh_shared.c`. To do this I tried using
   - [sigprocmask as a locking mechanism](differential/diff/41711139/)
   - [custom locking mechanism based on __sync_fetch_and_add](differential/diff/41711237/)

The "just wait" solution is too slow - some batches can take 20s to finish. In my testing (see test plan) I was not able to show the signal based approaches to be faster than the "naive" one implemented in this diff (which was surprising).

Here are example timings to cancel typechekcing job for
- this diff: {P58980434}
- sigprocmask version: {P58979757}
- sync_fetch_and_add version:{P58979827}

I can't see any improvement that would warrant introducing all the complexity for now. We can revisit if production data shows that things are too slow and A/B test the other ones, but I would like to start with this.

Reviewed By: gabelevi

Differential Revision: D6842414

fbshipit-source-id: d0a507c796b4a6919a4ef220f7d524f2eb79b307