Fix constness for CompactTaggedPtr::ptr(), remove arch-independent version
Summary:
The constness of the CompactTaggedPtr has nothing to do with the
constness of the pointee.

And since the bit-packing version appears to be supported on x86-64, ARM64, and
ppc64, delete the other version. Nothing is building or testing it, and it is
trivial to revive if we ever need it.

Reviewed By: ricklavoie

Differential Revision: D6853553

fbshipit-source-id: 9772cd153e51f6910ebf7766c17ae0e1857c32a4