Hack should PidLog.log the monitor PID, like Flow
Summary:
Currently the file /tmp/hh_server/zSdatazSuserszSljwzSwww-hg.pids or similar stores the PIDs of the server-main processes, and of all the workers. It is cleared out and first written to when the server main process starts up, and then appended each time a worker process is spawned.

The file is read only by "hh stop". It first tries to kill the existing hh_server nicely. If that doesn't work, then it kills it "meanly" by enumerating the PIDs file and killing each PID.

This diff makes it so the PIDs file also includes the monitor PID. The only effect for now is that "hh stop" will assuredly kill the monitor process. It was correctly terminating itself in the past, I believe, even in cases where we meanly killed the server, so this won't have any impact.

I'm making this change for a future diff in the stack. When the user clicks the "bug" button in their editor, and it sends a "telemetry/rage" request to the Hack LSP server, I want the LSP server to be able to fetch the "pstack" of every process. The most important process is the monitor. That's why it's important to keep track of the monitor PID.

I want to write out some facts about the way PidLog works to explain my choice of implementation...

1. The code calls "PidLog.init" from the start of server main. That opens a FD to the .pid file, which is "close_on_exec" so none of the spawned processes every get hold of it. But we need the monitor to write its PID even before server main has started.

2. One approach would have been to change it so that "PidLog.init" is called right at the very start of hh_server, as soon as we've parsed options to get root. And remove the close_on_exec flag, so that all sub-processes could happily write to it. But I felt kind of dodgy about doing that, about leaving an open FD just dangling amongst all the child processes.

3. Another approach would be to change it so "PidLog.init" merely deletes the .pid file and doesn't open any lasting FDs; all subsequent calls to "PidLog.log" would open/create the .pid file, append to it, then close it again. But this would require threading "root" through to every single spawn/exec in the entire codebase. I gave up on this after touching 15 files.

4. Another approach would be to invent a second file ".pid.monitor" whose sole job is to hold the PID of the monitor, and write to it like (3). This way it doesn't have creeping changes in the rest of the codebase, and remains small and self-contained. But it's a bit goofy.

5. The approach in this diff is one I copied from Flow. It merely records the monitor PID in the monitor, and passes that information to the server. It's clean and self-contained.

*** It has the downside that if the monitor has started but hasn't yet launched the server (e.g. because it's waiting for mercurial to settle) then it won't write the monitor's PID to the .pids file. That's a shame. But it seems like a small risk.

Reviewed By: gabelevi

Differential Revision: D6882185

fbshipit-source-id: 4a2811314d02a01ce37168b531a015ba1afb7e89