Better logging for monitor crash Unix.EINVAL/ECHILD
Summary:
Last week, two people reported that their hh just terminated after a rebase. Both of them had monitor logs like this:

[2018-02-01 17:22:33] check_and_run_loop_ threw with exception: Unix.Unix_error(Unix.EINVAL, "select", "")
[2018-02-01 17:22:34] typechecker exit with exit code 4
[2018-02-01 17:22:34] check_and_run_loop_ threw with exception: Unix.Unix_error(Unix.EINVAL, "select", "")
[2018-02-01 17:22:34] check_and_run_loop_ threw with Unix.ECHILD. Exiting

This diff attempts to add better logging (into the monitor log) with the stack trace that lead to EINVAL/ECHILD. I'd like to get to the bottom of why the monitor is terminating.

Reviewed By: dabek

Differential Revision: D6898166

fbshipit-source-id: 62869548acf2124111164bf07ac797ca47a71981