Change Hack LSP to return the signature of a hovered function, not its return type.
Summary:
Nuclide only shows the return type of a hovered function until Cmd is pressed, at which point it shows both the return type and the declaration.  Instead, we should always display the signature (including generic types), regardless of if Cmd is held or not.

Inferred generic types are displayed inline:
https://pxl.cl/bqff
https://pxl.cl/bqfd

Differential Revision: D6753898

fbshipit-source-id: 1f27a2a81f25faaf5b62a90be1fd00f319326c83