HHBBC shouldn't step through unreachable code while inserting RAT assertions
Summary:
While inserting RAT type assertions at the end of the optimize phase, HHBBC
shouldn't step through unreachable bytecode. There's an invariant that at this
point, the analysis of the function should have reached a fixed
point. Therefore, stepping through the bytecodes should not infer anything new
about the program. We do assert this in a few places (related to function
folding). However, unreachable parts of the function may not be analyzed fully,
and therefore stepping through them might break that invariant.

Instead, as soon as we hit an unreachable part of the program, just insert a
static analysis error guard and stop processing that block (same as we do
elsewhere).

Include a test case which triggers an assertion without this. The root cause of
why this test case triggers it is because of the lack of a pattern-match
optimization for IsType[C/L]; Not; Jmp[N]Z during analysis. The peephole
optimizer will remove the Not and flip the jump, but only during the
optimization phase. Add a suitable pattern matching optimization as well.

Reviewed By: markw65

Differential Revision: D6905679

fbshipit-source-id: f4e30540e589158216dc8117c742b340507f21f5