fix nastInitCheck for trait-protected members when constructor is overriden
Summary:
This is an edge case in nastInitCheck that actually does lead to broken behavior when you run the code we fail to report an error on. If a trait declares a constructor and initializes some members, it's possible that a class that uses the trait to override the constructor, in which case there is no way to recover the original trait constructor (like you could do with `parent::__construct()` for a traditional subclass), so the burden of initializing the trait-protected members falls on the using class.

This requires a small refactor in the way we compute `dc_deferred_init_members` for traits. We need to assume that traits never define a constructor and save all the members that could possibly need initialization later, and then decide when we actually have a `use` whether or not to merge them based on how the trait and the using class are defined. This does not affect checking the traits themselves because the only time we use `dc_deferred_init_members` is when we merge in parent/trait props.

The logic for when we fold in the deferred members for traits should be basically identical except in the "using class defines its own constructor" case.

Reviewed By: jamesjwu

Differential Revision: D6695001

fbshipit-source-id: 5af763952cabebb4e360927ced5e2e492a7b5bb5