Add nonnull type to Hack
Summary:
This diff is a first step towards adding `nonnull` type to Hack.  In particular, I ended up:

 - extending `Naming_special_names.Typehints` with a new value `"nonnull"`;
 - extending `Nast.hint_` with a new variant `Hnonnull`;
 - extending `Typing_defs.ty_` with a new variant `Tnonnull`.

Initially, I had to put lots of FIXMEs to make the code compile (exhaustiveness checks are amazing!).  I then resolved most of the FIXMEs one at a time (in TDD-kind-of-style) by first thinking of a test case that would exercise the relevant code, watching the test fail, resolving a FIXME, and watching the test pass (produce a reasonable error message or accept the code).  This is why this diff is largely tests and fixtures.  There are still FIXMEs left in the code for which I wasn't able to come up with code examples that would hit the corresponding code paths.  I hope we can discuss those.  Also, some of the FIXMEs looked obvious enough to be resolved without tests.

Reviewed By: andrewjkennedy

Differential Revision: D6883613

fbshipit-source-id: e79b3eae34265ed7619ce85153f393b13669081a