slightly relax brutal treatment of FPassL
Summary:
The treatment of FPassL introduced in D6358985 wipes out all
equalities involving variables that might be passed by reference to
functions, including the one that relates the actual variables
passed. That turns out to break at least one test case. This diff
leaves the equality between the variables passed intact.

It is not *absolutely* clear to me that this is sound in the general
case. A counterexample would involve a called function that created
some aliases and then some code in the caller that used the aliases to
silently break the equality. I can't immediately construct such an
example that would still get through semdiff, though, and nor can I
see a clearly sound fix for the original test failure that isn't
ridiculously complicated. So I'm going to "move fast" by checking this
in and seeing if anything breaks later. (It'd be nice if we could
trust the hints.)

Reviewed By: vladima

Differential Revision: D6937504

fbshipit-source-id: 90dc8acfd69b5e7345c88e9e8488e5415a305e63