Add separate cmake file for setting OS variables
In the Options.cmake file, MAP_TEXT_HUGE_PAGES and ENABLE_FOLLY_SYMBOLIZER
are set depending on the value of the LINUX variable, which is only set in
a file included later - HPHPCompiler.cmake. This causes these 2 variables not
to be visible in third-party modules the first time configure is run
(subsequent times the variables are taken from the cmake cache).

This change fixes this behavior by adding a new cmake file that sets the OS
variables and including it before Options.cmake.