file_mode fix
Summary:
`hh_server` uses the `option`-part of the `Parse_result.fi_mode` to determine
whether a file should or should not be added to the `Decl_heap`. During the
development of the FFP / lowerer, we introduced an explicit `FileInfo.Mphp`. By
using `Some Mphp` instead of `None`, `hh_server` quietly started taking PHP
files in as if Hack files.

Reviewed By: andrewjkennedy

Differential Revision: D6950049

fbshipit-source-id: 6afd80b19668fbe58145bbf557101619a4f0e54e