Add support for XED-generated padding instructions
Add XED-generated int3(n) and nop(n).
X64Assembler::gs(),fs() methods removed. This functionality is
moved to the MemoryRef class which now has MemoryRef::segment, used
by the XedOperand constructor to generate a xed_encoder_operand_t
value in which the default segment is replaced with the one specified
in the MemoryRef::segment member. The prefix(...) function in
vasm-x64.cpp is now a template and can be used with a MemoryRef
argument when HAVE_LIBXED is defined or a X64Assembler argument
otherwise (both of them feature the gs(),fs() interface).
Fix warnings caused by variables used only for asserts in the xedEmitN
methods and make their code slightly more readable by moving the
final assert to a reusable macro.