find_hh script should retry on client disconnection during ClientConnect
Summary:
A very common use case when developing is stringing together
commands like

$ hg update master && hh . && arc rebuild

or similar.

This doesn't work well when the server crashes suddenly or (more recently)
when the server is intentionally killed by the Informant to load a new
saved state, since the "hh ." command will fail while it's "Processing \" spinner
is showing.

The spinner runs during ClientConnect code, and so a sudden disconnect would
cause Exit_status.No_server_running, which results in a return code of 6.

To fix the above problem, we retry the entire hh script on this return code.

We must retry the entire script instead of just repeating the hh_client
invocation to avoid version mismatch battles (see comment in diff).

Note: This retry isn't exercised by "hh lsp" or "hh build" because those catch the No_server_running exception and never exit with code 6.

See newly added docs in code for more details.

Reviewed By: ljw1004

Differential Revision: D6867698

fbshipit-source-id: a964db68464ebd8c4cf58918737259d11f4ab042