Experimental option to reject lambdas that cannot be type checked at definition site
Summary:
If the `disallow_ambiguous_lambda` option is set in .hhconfig then any lambda that cannot
be checked at definition site will be flagged with an error. The error gives an indication
of how many times the body was checked at the use-site.

Also we extend the range of "untyped" contexts in which lambdas are checked at definition-site
with untyped parameters.

Reviewed By: dlreeves

Differential Revision: D6871587

fbshipit-source-id: ef0b3fd135259d43de09dd42b2fe630691f89415