Ensure that hhbbc emits contiguous fpi regions
Summary:
If hhbbc can prove that a particular path leaves the fpi region via an
exception, but the FCall* is still reachable, reverse post order
doesn't guarantee that the fpi region is contiguous.

This adds fake edges from any terminal blocks in the fpi region to the
corresponding FCall block (if there is one) to ensure that rpo does
make the region contiguous.

Reviewed By: ricklavoie

Differential Revision: D6959684

fbshipit-source-id: f5d00cedd65be5355a5fa1d4454bb37212d25df7