Avoid generating Jmps around Jmp[N]Zs in hhbbc
Summary:
If a block ends with a JmpZ or JmpNZ, and the emitter picks a block
ordering that places the target immediately after the Jmp, we'll end
up spitting out a Jmp[N]Z followed by a Jmp to the fallthrough,
followed by the target block.

This flips the sense of the Jmp[N]Z to avoid the extra Jmp.

Reviewed By: ricklavoie

Differential Revision: D6960099

fbshipit-source-id: db24e2ad90ac4a447f7d60fd229f36ed8a36704a