Add option to disable parameter type compatibility checks
Summary: hh_server can already do a more sophisticated version of this check (properly accounting for type aliases, and class co/contra-variance) which the simple string comparison in hhvm can't handle. An audit of hhbbc suggests we don't actually use this particular invariant anywhere, and enforcing it in the same fashion as hack would require changes to the way autoloading is observed when classes are loaded. Since it isn't adding much value add an option to suppress it.

Reviewed By: markw65

Differential Revision: D6963404

fbshipit-source-id: 0845057e64ea6cd0331feda3291cd66a64127879