Make ReflectionProperty implementation suck less.
Summary:
ReflectionProperty construction involved calling a helper that set native data _and_ returned an array of info (or negative integers, in which case the constructor was supposed to throw). The native data was luckily never read, because it could not be used safely: it could be set to one of 2 different classes with no way to tell them apart. The array of info was set in a public property (which was not part of the API), and used to implement all the functionality of the class.

Fix this by folding the helper functionality into the constructor, making the native data actually usable, and using it to implement all the low level functionality of the class with native methods. Fix a long-standing bug where ->class and ->getDeclaringClass() produced different results when a property was inherited (unlike in PHP5 and PHP7 https://3v4l.org/AUqCs ). Add a separate unit test for it and fix a test expectation that was verifying the broken behavior. Add missing tests for the constructor throwing.

I did not remove the now unused, public-but-not-in-the-API info property in this diff because it has a big impact on test expectations, and I want to show that this diff does not change behavior aside from the one bugfix. It will be removed in the next diff.

Reviewed By: markw65

Differential Revision: D6958622

fbshipit-source-id: 6a0cf689ea8b4c735a62706e3584579b2fc7ed2d